Extract BERT into skainet-bert and shared tokenizer into skainet-llm
Move shared tokenizer abstractions (Tokenizer, TokenizerType, TokenizerStrategy,
and all strategy implementations) from skainet-kllama into a new skainet-llm
module, and extract all BERT code into a new skainet-bert module. This decouples
BERT from kllama so each can evolve independently while sharing tokenizer types
through skainet-llm.

- New module skainet-apps/skainet-llm (sk.ainet.apps.llm)
- New module skainet-apps/skainet-bert (sk.ainet.apps.bert)
- Update skainet-kllama to depend on skainet-llm, fix imports
- Update skainet-kbert-cli to depend on skainet-bert, move Main.kt
