+++ settings.gradle.kts
@@ -72,6 +72,9 @@ include("skainet-test:skainet-test-groundtruth")
 // ====== APPS
 include("skainet-apps:skainet-grayscale-cli")
 include("skainet-apps:skainet-tensor-tools")
+include("skainet-apps:skainet-llm")
+include("skainet-apps:skainet-bert")
 include("skainet-apps:skainet-kllama")
 include("skainet-apps:skainet-kllama-cli")
+include("skainet-apps:skainet-kbert-cli")
 include("skainet-io:skainet-io-safetensors")
