+++ skainet-apps/skainet-bert/src/jvmTest/kotlin/sk/ainet/apps/bert/BertNumericalAccuracyTest.kt
@@ -1,0 +1,365 @@
+package sk.ainet.apps.bert
+
+import kotlinx.coroutines.test.runTest
+import org.junit.Assume.assumeTrue
+import sk.ainet.context.DirectCpuExecutionContext
+import sk.ainet.io.JvmRandomAccessSource
+import sk.ainet.io.safetensors.SafeTensorsParametersLoader
+import sk.ainet.lang.tensor.Tensor
+import sk.ainet.lang.tensor.data.FloatArrayTensorData
+import sk.ainet.lang.types.DType
+import sk.ainet.lang.types.FP32
+import java.nio.file.Path
+import kotlin.io.path.exists
+import kotlin.io.path.readText
+import kotlin.math.abs
+import kotlin.math.sqrt
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertTrue
+
+/**
+ * Numerical accuracy test for BERT implementation against PyTorch/sentence-transformers.
+ *
+ * Requires:
+ * 1. Model cached at ~/.cache/huggingface/hub/models--MongoDB--mdbr-leaf-ir/
+ * 2. Reference data generated by bert-research/generate_reference.py
+ *
+ * Tests are skipped (via JUnit Assume) when either is unavailable.
+ */
+class BertNumericalAccuracyTest {
+
+    companion object {
+        private const val MODEL_ID = "MongoDB--mdbr-leaf-ir"
+        /** Absolute tolerance for float32 comparison. Slightly relaxed from 1e-4
+         *  to accommodate rounding accumulation across 6 encoder layers. */
+        private const val TOLERANCE = 5e-4
+
+        /** Resolve HuggingFace cache snapshot directory for a model. */
+        private fun resolveHfCacheDir(): Path? {
+            val home = System.getProperty("user.home") ?: return null
+            val modelsDir = Path.of(home, ".cache", "huggingface", "hub", "models--$MODEL_ID", "snapshots")
+            if (!modelsDir.exists()) return null
+            // Use first (usually only) snapshot
+            val snapshot = modelsDir.toFile().listFiles()?.firstOrNull { it.isDirectory } ?: return null
+            return snapshot.toPath()
+        }
+
+        /** Resolve reference_data.json by walking up from working directory. */
+        private fun resolveReferenceDataPath(): Path {
+            var dir = Path.of("").toAbsolutePath()
+            // Walk up until we find bert-research/reference_data.json or reach root
+            while (dir.parent != null) {
+                val candidate = dir.resolve("bert-research/reference_data.json")
+                if (candidate.exists()) return candidate
+                dir = dir.parent
+            }
+            // Fallback: return the original path (will be caught by assumeTrue)
+            return Path.of("").toAbsolutePath().resolve("bert-research/reference_data.json")
+        }
+    }
+
+    private val ctx = DirectCpuExecutionContext()
+
+    // =========================================================================
+    // JSON parsing helpers (minimal, avoids external dependency)
+    // =========================================================================
+
+    private data class ReferenceData(
+        val inputs: List<InputReference>,
+        val cosineSimilarity: Double
+    )
+
+    private data class InputReference(
+        val text: String,
+        val inputIds: List<Int>,
+        val attentionMask: List<Int>,
+        val tokenTypeIds: List<Int>,
+        val hiddenStates: List<List<Double>>,
+        val embedding: List<Double>
+    )
+
+    private fun parseReferenceJson(json: String): ReferenceData {
+        val inputs = mutableListOf<InputReference>()
+        // Extract cosine_similarity
+        val simMatch = Regex("\"cosine_similarity\"\\s*:\\s*([\\d.eE\\-+]+)").find(json)
+        val cosineSim = simMatch?.groupValues?.get(1)?.toDoubleOrNull() ?: error("Missing cosine_similarity")
+
+        // Extract each input block by finding "text": patterns
+        val inputPattern = Regex("\\{\\s*\"text\"\\s*:")
+        val inputStarts = inputPattern.findAll(json).map { it.range.first }.toList()
+
+        for ((idx, start) in inputStarts.withIndex()) {
+            val end = if (idx + 1 < inputStarts.size) {
+                // Find the closing brace of this entry before the next "text" key
+                findMatchingBrace(json, start)
+            } else {
+                findMatchingBrace(json, start)
+            }
+            val block = json.substring(start, end + 1)
+            inputs.add(parseInputBlock(block))
+        }
+
+        return ReferenceData(inputs, cosineSim)
+    }
+
+    private fun findMatchingBrace(json: String, openPos: Int): Int {
+        var depth = 0
+        var inString = false
+        var escape = false
+        for (i in openPos until json.length) {
+            val c = json[i]
+            if (escape) { escape = false; continue }
+            if (c == '\\' && inString) { escape = true; continue }
+            if (c == '"' && !escape) { inString = !inString; continue }
+            if (inString) continue
+            if (c == '{') depth++
+            if (c == '}') { depth--; if (depth == 0) return i }
+        }
+        error("Unmatched brace at position $openPos")
+    }
+
+    private fun parseInputBlock(block: String): InputReference {
+        val text = extractString(block, "text")
+        val inputIds = extractIntArray(block, "input_ids")
+        val attentionMask = extractIntArray(block, "attention_mask")
+        val tokenTypeIds = extractIntArray(block, "token_type_ids")
+        val hiddenStates = extractNestedArray(block, "hidden_states")
+        val embedding = extractFlatArray(block, "embedding")
+        return InputReference(text, inputIds, attentionMask, tokenTypeIds, hiddenStates, embedding)
+    }
+
+    private fun extractString(json: String, key: String): String {
+        val pattern = Regex("\"$key\"\\s*:\\s*\"([^\"]+)\"")
+        return pattern.find(json)?.groupValues?.get(1) ?: error("Missing key: $key")
+    }
+
+    private fun extractIntArray(json: String, key: String): List<Int> {
+        val keyIdx = json.indexOf("\"$key\"")
+        if (keyIdx < 0) error("Missing key: $key")
+        val arrStart = json.indexOf('[', keyIdx)
+        val arrEnd = json.indexOf(']', arrStart)
+        val content = json.substring(arrStart + 1, arrEnd).trim()
+        if (content.isEmpty()) return emptyList()
+        return content.split(",").map { it.trim().toInt() }
+    }
+
+    private fun extractFlatArray(json: String, key: String): List<Double> {
+        val keyIdx = json.indexOf("\"$key\"")
+        if (keyIdx < 0) error("Missing key: $key")
+        val arrStart = json.indexOf('[', keyIdx)
+        val arrEnd = findMatchingBracket(json, arrStart)
+        val content = json.substring(arrStart + 1, arrEnd).trim()
+        if (content.isEmpty()) return emptyList()
+        return content.split(",").map { it.trim().toDouble() }
+    }
+
+    private fun extractNestedArray(json: String, key: String): List<List<Double>> {
+        val keyIdx = json.indexOf("\"$key\"")
+        if (keyIdx < 0) error("Missing key: $key")
+        val outerStart = json.indexOf('[', keyIdx)
+        val outerEnd = findMatchingBracket(json, outerStart)
+        val outerContent = json.substring(outerStart + 1, outerEnd)
+
+        val rows = mutableListOf<List<Double>>()
+        var pos = 0
+        while (pos < outerContent.length) {
+            val rowStart = outerContent.indexOf('[', pos)
+            if (rowStart < 0) break
+            val rowEnd = outerContent.indexOf(']', rowStart)
+            val rowContent = outerContent.substring(rowStart + 1, rowEnd).trim()
+            if (rowContent.isNotEmpty()) {
+                rows.add(rowContent.split(",").map { it.trim().toDouble() })
+            }
+            pos = rowEnd + 1
+        }
+        return rows
+    }
+
+    private fun findMatchingBracket(json: String, openPos: Int): Int {
+        var depth = 0
+        for (i in openPos until json.length) {
+            if (json[i] == '[') depth++
+            if (json[i] == ']') { depth--; if (depth == 0) return i }
+        }
+        error("Unmatched bracket at position $openPos")
+    }
+
+    // =========================================================================
+    // Tensor helpers
+    // =========================================================================
+
+    private fun <T : DType> Tensor<T, Float>.toFloatList(): List<Float> {
+        val data = this.data
+        if (data is FloatArrayTensorData<*>) return data.buffer.toList()
+        return data.copyToFloatArray().toList()
+    }
+
+    private fun cosineSimilarity(a: List<Float>, b: List<Float>): Float {
+        require(a.size == b.size)
+        var dot = 0f; var nA = 0f; var nB = 0f
+        for (i in a.indices) {
+            dot += a[i] * b[i]
+            nA += a[i] * a[i]
+            nB += b[i] * b[i]
+        }
+        val denom = sqrt(nA) * sqrt(nB)
+        return if (denom > 0f) dot / denom else 0f
+    }
+
+    // =========================================================================
+    // Tests
+    // =========================================================================
+
+    @Test
+    fun tokenizer_matchesPythonTokenIds() {
+        val refPath = resolveReferenceDataPath()
+        assumeTrue("Reference data not found at $refPath", refPath.exists())
+        val modelDir = resolveHfCacheDir()
+        assumeTrue("Model not cached", modelDir != null)
+        modelDir!!
+
+        val vocabPath = modelDir.resolve("vocab.txt")
+        assumeTrue("vocab.txt not found", vocabPath.exists())
+
+        val tokenizer = HuggingFaceTokenizer.fromVocabTxt(vocabPath.readText())
+        val reference = parseReferenceJson(refPath.readText())
+
+        for (input in reference.inputs) {
+            val output = tokenizer.encodeWithMetadata(input.text)
+            assertEquals(
+                input.inputIds,
+                output.inputIds.toList(),
+                "Token IDs mismatch for \"${input.text}\""
+            )
+            assertEquals(
+                input.attentionMask,
+                output.attentionMask.toList(),
+                "Attention mask mismatch for \"${input.text}\""
+            )
+            assertEquals(
+                input.tokenTypeIds,
+                output.tokenTypeIds.toList(),
+                "Token type IDs mismatch for \"${input.text}\""
+            )
+        }
+    }
+
+    @Test
+    fun forward_matchesPythonHiddenStates() = runTest {
+        val refPath = resolveReferenceDataPath()
+        assumeTrue("Reference data not found at $refPath", refPath.exists())
+        val modelDir = resolveHfCacheDir()
+        assumeTrue("Model not cached", modelDir != null)
+        modelDir!!
+
+        val mainSafetensors = modelDir.resolve("model.safetensors")
+        assumeTrue("model.safetensors not found", mainSafetensors.exists())
+
+        val config = MDBR_LEAF_IR_CONFIG
+        val reference = parseReferenceJson(refPath.readText())
+
+        // Load model without projection (for forward() we only need the encoder)
+        val encoderConfig = config.copy(projectionDim = null)
+        val loader = SafeTensorsParametersLoader(
+            sourceProvider = { JvmRandomAccessSource.open(mainSafetensors.toString()) }
+        )
+        val weights = loadBertWeights(loader, ctx, FP32::class, encoderConfig)
+        val runtime = BertRuntime(ctx, weights, FP32::class)
+
+        for (input in reference.inputs) {
+            val tokenIds = input.inputIds.toIntArray()
+            val hiddenStates = runtime.forward(tokenIds)
+            val actual = hiddenStates.toFloatList()
+
+            val expectedFlat = input.hiddenStates.flatten()
+            assertEquals(
+                expectedFlat.size, actual.size,
+                "Hidden states size mismatch for \"${input.text}\""
+            )
+
+            var maxDiff = 0.0
+            for (i in actual.indices) {
+                val diff = abs(actual[i].toDouble() - expectedFlat[i])
+                if (diff > maxDiff) maxDiff = diff
+            }
+
+            assertTrue(
+                maxDiff < TOLERANCE,
+                "Hidden states max diff $maxDiff exceeds tolerance $TOLERANCE for \"${input.text}\""
+            )
+            println("forward(\"${input.text}\"): max diff = $maxDiff (tolerance = $TOLERANCE)")
+        }
+    }
+
+    @Test
+    fun encode_matchesPythonEmbedding() = runTest {
+        val refPath = resolveReferenceDataPath()
+        assumeTrue("Reference data not found at $refPath", refPath.exists())
+        val modelDir = resolveHfCacheDir()
+        assumeTrue("Model not cached", modelDir != null)
+        modelDir!!
+
+        val mainSafetensors = modelDir.resolve("model.safetensors")
+        val denseSafetensors = modelDir.resolve("2_Dense").resolve("model.safetensors")
+        assumeTrue("model.safetensors not found", mainSafetensors.exists())
+        assumeTrue("2_Dense/model.safetensors not found", denseSafetensors.exists())
+
+        val config = MDBR_LEAF_IR_CONFIG
+        val reference = parseReferenceJson(refPath.readText())
+
+        // Load model with both encoder and projection weights
+        val loaders = listOf(
+            SafeTensorsParametersLoader(
+                sourceProvider = { JvmRandomAccessSource.open(mainSafetensors.toString()) }
+            ),
+            SafeTensorsParametersLoader(
+                sourceProvider = { JvmRandomAccessSource.open(denseSafetensors.toString()) }
+            )
+        )
+        val weights = loadBertWeights(loaders, ctx, FP32::class, config)
+        val runtime = BertRuntime(ctx, weights, FP32::class)
+
+        val embeddings = mutableListOf<List<Float>>()
+
+        for (input in reference.inputs) {
+            val tokenIds = input.inputIds.toIntArray()
+            val attentionMask = input.attentionMask.toIntArray()
+            val tokenTypeIds = input.tokenTypeIds.toIntArray()
+
+            val embedding = runtime.encode(tokenIds, attentionMask, tokenTypeIds)
+            val actual = embedding.toFloatList()
+            embeddings.add(actual)
+
+            assertEquals(
+                input.embedding.size, actual.size,
+                "Embedding size mismatch for \"${input.text}\""
+            )
+
+            var maxDiff = 0.0
+            for (i in actual.indices) {
+                val diff = abs(actual[i].toDouble() - input.embedding[i])
+                if (diff > maxDiff) maxDiff = diff
+            }
+
+            assertTrue(
+                maxDiff < TOLERANCE,
+                "Embedding max diff $maxDiff exceeds tolerance $TOLERANCE for \"${input.text}\""
+            )
+            println("encode(\"${input.text}\"): max diff = $maxDiff (tolerance = $TOLERANCE)")
+        }
+
+        // Compare cosine similarity
+        if (embeddings.size >= 2) {
+            val sim = cosineSimilarity(embeddings[0], embeddings[1]).toDouble()
+            val expectedSim = reference.cosineSimilarity
+            val simDiff = abs(sim - expectedSim)
+            assertTrue(
+                simDiff < TOLERANCE,
+                "Cosine similarity diff $simDiff exceeds tolerance $TOLERANCE " +
+                    "(got $sim, expected $expectedSim)"
+            )
+            println("Cosine similarity: $sim (expected $expectedSim, diff = $simDiff)")
+        }
+    }
+}
