--- skainet-apps/skainet-kllama/src/commonMain/kotlin/sk/ainet/apps/kllama/Tokenizer.kt
@@ -1,7 +1,0 @@
-package sk.ainet.apps.kllama
-
-interface Tokenizer {
-    fun encode(text: String): IntArray
-    fun decode(tokens: IntArray): String
-    fun decode(token: Int): String
-}
