+++ skainet-apps/skainet-llm/src/commonMain/kotlin/sk/ainet/apps/llm/Tokenizer.kt
@@ -1,0 +1,7 @@
+package sk.ainet.apps.llm
+
+interface Tokenizer {
+    fun encode(text: String): IntArray
+    fun decode(tokens: IntArray): String
+    fun decode(token: Int): String
+}
